make separate github service for github api calls
