+++ internal/app/repository/domain.go
@@ -2,21 +2,6 @@ package repository
 
 import "time"
 
-type RepoOwner struct {
-	Login string `json:"login"`
-}
-
-type FetchRepositoryDetailsResponse struct {
-	Id              int       `json:"id"`
-	Name            string    `json:"name"`
-	Description     string    `json:"description"`
-	LanguagesURL    string    `json:"languages_url"`
-	UpdateDate      time.Time `json:"updated_at"`
-	RepoOwnerName   RepoOwner `json:"owner"`
-	ContributorsUrl string    `json:"contributors_url"`
-	RepoUrl         string    `json:"html_url"`
-}
-
 type Repository struct {
 	Id              int
 	GithubRepoId    int
@@ -37,19 +22,6 @@ type FetchUsersContributedReposResponse struct {
 	Repository
 	Languages        []string
 	TotalCoinsEarned int
-}
-
-type FetchRepoContributorsResponse struct {
-	Id            int    `json:"id"`
-	Name          string `json:"login"`
-	AvatarUrl     string `json:"avatar_url"`
-	GithubUrl     string `json:"html_url"`
-	Contributions int    `json:"contributions"`
-}
-
-type FetchParticularRepoDetails struct {
-	Repository
-	Contributors []FetchRepoContributorsResponse
 }
 
 type Contribution struct {
